pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-taiga-bridge"
version = "0.1.0"
description = "Taiga integration bridge for MCP"
requires-python = ">=3.10"
authors = [{ name = "Taiga MCP Bridge Team" }]
dependencies = [
    "mcp[cli]>=0.1.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.23.0",
    "python-taiga>=0.9.0",
    "python-dotenv>=0.15.0",
    "httpx>=0.25.0",
    "pydantic>=2.4.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "responses>=0.23.1",
    "black>=23.3.0",
    "isort>=5.12.0",
    "mypy>=1.3.0",
    "flake8>=6.0.0",
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[tool.setuptools]
packages = ["src"]

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line-length = 100
target-version = ['py310']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: Run unit tests",
    "integration: Run integration tests",
    "slow: Mark tests as slow-running",
    "auth: Authentication tests",
    "core: Core functionality tests",
    "projects: Project-related tests",
    "epics: Epic-related tests",
    "user_stories: User story-related tests",
    "tasks: Task-related tests",
    "issues: Issue-related tests",
    "sprints: Sprint-related tests",
]
addopts = "-v"
